requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
solana==0.30.2
solders==0.18.1
base58==2.1.1
anchorpy==0.18.0
# Metaplex integration via Node.js (see metaplex/ directory)
requests==2.31.0
Pillow==10.1.0

